Full conversion of addon to webextension
- fixes #3382
- fixes #3183
- fixes #3176
- fixes #3177
- fixes #3182
- fixes #3180
- fixes #3313
