+++ addon/icons/icon-32.png
