+++ addon/icons/icon-96.png
