+++ addon/icons/notification-icon.png
