--- addon/lib/webpack-after-build.js
@@ -1,11 +1,0 @@
-module.exports = class AfterBuildPlugin {
-  constructor (cb) {
-    this.cb = cb;
-  }
-
-  apply (compiler) {
-    compiler.plugin('done', () => {
-      this.cb();
-    });
-  }
-};
