--- addon/src/chrome.manifest
@@ -1,1 +1,0 @@
-content testpilot-addon-icons ./webextension/icons/
