--- addon/src/install.rdf
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RDF xmlns="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:em="http://www.mozilla.org/2004/em-rdf#">
-  <Description about="urn:mozilla:install-manifest">
-    <em:id>@testpilot-addon</em:id>
-    <em:type>2</em:type>
-    <em:bootstrap>true</em:bootstrap>
-    <em:hasEmbeddedWebExtension>true</em:hasEmbeddedWebExtension>
-    <em:unpack>false</em:unpack>
-    <em:version>2.0.3</em:version>
-    <em:name>Test Pilot</em:name>
-    <em:description>Test Pilot is a privacy-sensitive user research program focused on getting new features into Firefox faster.</em:description>
-    <em:creator>Mozilla (https://mozilla.org/)</em:creator>
-    <em:homepageURL>https://testpilot.firefox.com/</em:homepageURL>
-    <em:updateURL>https://testpilot.firefox.com/static/addon/update.rdf</em:updateURL>
-    <em:iconURL>chrome://testpilot-addon-icons/content/icon-96.png</em:iconURL>
-    <em:multiprocessCompatible>true</em:multiprocessCompatible>
-    <em:targetApplication>
-      <!-- Firefox -->
-      <Description>
-        <em:id>{ec8030f7-c20a-464f-9b0e-13a3a9e97384}</em:id>
-        <em:minVersion>58.0</em:minVersion>
-        <em:maxVersion>*</em:maxVersion>
-      </Description>
-    </em:targetApplication>
-  </Description>
-</RDF>
