--- addon/test/test-setup.js
@@ -1,1 +1,0 @@
-// Global setup for all tests
